refactor: 删除无用代码
