+++ lark/handlers/common.go
@@ -3,10 +3,7 @@ package handlers
 import (
 	"encoding/json"
 	"fmt"
-	"lark/services"
 	"regexp"
-	"strconv"
-	"strings"
 )
 
 // func sendCard
@@ -26,64 +23,4 @@ func parseContent(content string) string {
 	}
 	text := contentMap["text"].(string)
 	return msgFilter(text)
-}
-func processMessage(msg interface{}) (string, error) {
-	msg = strings.TrimSpace(msg.(string))
-	msgB, err := json.Marshal(msg)
-	if err != nil {
-		return "", err
-	}
-
-	msgStr := string(msgB)
-
-	if len(msgStr) >= 2 {
-		msgStr = msgStr[1 : len(msgStr)-1]
-	}
-	return msgStr, nil
-}
-
-func processNewLine(msg string) string {
-	return strings.Replace(msg, "\\n", `
-`, -1)
-}
-
-func processQuote(msg string) string {
-	return strings.Replace(msg, "\\\"", "\"", -1)
-}
-
-// 将字符中 \u003c 替换为 <  等等
-func processUnicode(msg string) string {
-	regex := regexp.MustCompile(`\\u[0-9a-fA-F]{4}`)
-	return regex.ReplaceAllStringFunc(msg, func(s string) string {
-		r, _ := regexp.Compile(`\\u`)
-		s = r.ReplaceAllString(s, "")
-		i, _ := strconv.ParseInt(s, 16, 32)
-		return string(i)
-	})
-}
-
-func cleanTextBlock(msg string) string {
-	msg = processNewLine(msg)
-	msg = processUnicode(msg)
-	msg = processQuote(msg)
-	return msg
-}
-
-func limitContextLen(msg []services.Messages) []services.Messages {
-	maxLength := 4096
-	//限制对话上下文长度
-	for getStrPoolTotalLength(msg) > maxLength {
-		msg = append(msg[:1], msg[2:]...)
-	}
-
-	return msg
-}
-
-func getStrPoolTotalLength(strPool []services.Messages) int {
-	var total int
-	for _, v := range strPool {
-		bytes, _ := json.Marshal(v)
-		total += len(string(bytes))
-	}
-	return total
 }
