+++ lark/handlers/discord.go
@@ -45,22 +45,9 @@ type ReqCb struct {
 type Scene string
 
 const (
-	/**
-	 * 首次触发生成
-	 */
-	FirstTrigger Scene = "FirstTrigger"
-	/**
-	 * 生成图片结束
-	 */
-	GenerateEnd Scene = "GenerateEnd"
-	/**
-	 * 发送的指令midjourney生成过程中发现错误
-	 */
+	FirstTrigger      Scene = "FirstTrigger"
+	GenerateEnd       Scene = "GenerateEnd"
 	GenerateEditError Scene = "GenerateEditError"
-	/**
-	 * 发送的指令midjourney直接报错或排队阻塞不在该项目中处理 在业务服务中处理
-	 * 例如：首次触发生成多少秒后没有回调业务服务判定会指令错误或者排队阻塞
-	 */
 )
 
 func DiscordHandler(c *gin.Context) {
@@ -87,7 +74,13 @@ func DiscordHandler(c *gin.Context) {
 			if params.Discord.MessageReference != nil {
 				referenceMsgId = params.Discord.MessageReference.MessageID
 			}
-			discordTriggerReplayLark(params.Discord.Attachments[0].URL, params.Discord.Message.ID, id, msgHash, referenceMsgId)
+			discordTriggerReplayLark(
+				params.Discord.Attachments[0].URL,
+				params.Discord.Message.ID,
+				id,
+				msgHash,
+				referenceMsgId,
+			)
 		}
 		return
 	}
@@ -95,7 +88,6 @@ func DiscordHandler(c *gin.Context) {
 	if params.Type == GenerateEditError {
 		if id, notFound := getDiscordLardMapId(params.Content); notFound == nil {
 			if idl, err := getDiscordLarkMapJson(id); err == nil {
-				//TODO 并未考虑 升级过程中导致的Stopped情况
 				instructException(id, idl.LarkChatId, idl.MsgId)
 			}
 		}
@@ -104,16 +96,11 @@ func DiscordHandler(c *gin.Context) {
 }
 
 func SendDiscordMessageBot(msgId string, content string, ctx context.Context, larkChatId string) {
-
-	// 检查排队
 	err := discordQueueCheck(larkChatId)
 	if err != nil {
 		chore.ReplyMsg(ctx, err.Error(), &msgId)
 		return
 	}
-
-	fmt.Println("排队检查结束")
-
 	str := msgId + strconv.FormatInt(time.Now().UnixNano(), 10)
 	hash := md5.Sum([]byte(str))
 	id := hex.EncodeToString(hash[:])[:12]
@@ -123,12 +110,10 @@ func SendDiscordMessageBot(msgId string, content string, ctx context.Context, la
 		LarkMsgIdMapPrevDiscordMsgId: map[string]string{},
 		LarkChatId:                   larkChatId,
 	})
-	fmt.Println("准备请求midjourney")
 	err = services.ReqMidjourney(services.RequestTrigger{
 		Type:   "generate",
 		Prompt: DiscordPrefix + id + DiscordNextFix + content,
 	})
-	fmt.Println("请求midjourney结束")
 	if err != nil {
 		db.GetCache().Clear(id)
 		chore.ReplyMsg(ctx, fmt.Sprintf("🤖️：图片生成失败，请稍后再试～\n错误信息: %v", err), &msgId)
@@ -145,11 +130,20 @@ func SendDiscordMessageBot(msgId string, content string, ctx context.Context, la
 	checkSendDiscordMessage(make(chan struct{}), id)
 }
 
-func SendDiscordMessageBotUpscale(index int64, discordMessageId string, msgHash string, larkMsgId string) error {
+func SendDiscordMessageBotUpscale(
+	index int64,
+	discordMessageId string,
+	msgHash string,
+	larkMsgId string,
+) error {
 	/******* 处理同一张图片 点击同一个事件 start *******/
 	flagStr := db.GetCache().Get(discordMessageId)
 	if strings.Contains(flagStr, fmt.Sprint("U", index)) {
-		chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：您已经给该照片升级过: %v", fmt.Sprint("U", index)), &larkMsgId)
+		chore.ReplyMsg(
+			context.Background(),
+			fmt.Sprintf("🤖️：您已经给该照片升级过: %v", fmt.Sprint("U", index)),
+			&larkMsgId,
+		)
 		return errors.New("已经升级过")
 	}
 	db.GetCache().Set(discordMessageId, flagStr+fmt.Sprint("U", index))
@@ -162,10 +156,18 @@ func SendDiscordMessageBotUpscale(index int64, discordMessageId string, msgHash 
 		Index:        index,
 	})
 	if err != nil {
-		chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：图片升级失败，请稍后再试～\n错误信息: %v", err), &larkMsgId)
+		chore.ReplyMsg(
+			context.Background(),
+			fmt.Sprintf("🤖️：图片升级失败，请稍后再试～\n错误信息: %v", err),
+			&larkMsgId,
+		)
 		return err
 	}
-	chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：图片正在进行%v操作升级，请稍等......", fmt.Sprint("U", index)), &larkMsgId)
+	chore.ReplyMsg(
+		context.Background(),
+		fmt.Sprintf("🤖️：图片正在进行%v操作升级，请稍等......", fmt.Sprint("U", index)),
+		&larkMsgId,
+	)
 	return nil
 }
 
@@ -177,10 +179,19 @@ func SendDiscordMessageBotV(index int64, discordMessageId string, msgHash string
 		Index:        index,
 	})
 	if err != nil {
-		chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：图片操作失败，请稍后再试~\n错误信息: %v", err), &larkMsgId)
+		chore.ReplyMsg(
+			context.Background(),
+			fmt.Sprintf("🤖️：图片操作失败，请稍后再试~\n错误信息: %v", err),
+			&larkMsgId,
+		)
 		return err
 	}
-	chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：图片正在进行%v操作，请稍等......", fmt.Sprint("V", index)), &larkMsgId)
+	chore.ReplyMsg(
+		context.Background(),
+		fmt.Sprintf("🤖️：图片正在进行%v操作，请稍等......",
+			fmt.Sprint("V", index)),
+		&larkMsgId,
+	)
 	return nil
 }
 
@@ -191,10 +202,18 @@ func SendDiscordMessageMaxUpscale(discordMessageId string, msgHash string, larkM
 		MsgHash:      msgHash,
 	})
 	if err != nil {
-		chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：图片升级失败，请稍后再试～\n错误信息: %v", err), &larkMsgId)
+		chore.ReplyMsg(
+			context.Background(),
+			fmt.Sprintf("🤖️：图片升级失败，请稍后再试～\n错误信息: %v", err),
+			&larkMsgId,
+		)
 		return err
 	}
-	chore.ReplyMsg(context.Background(), "🤖️：图片正在进行最大升级，请稍等......", &larkMsgId)
+	chore.ReplyMsg(
+		context.Background(),
+		"🤖️：图片正在进行最终升级，请稍等......",
+		&larkMsgId,
+	)
 	return nil
 }
 
@@ -205,10 +224,18 @@ func SendDiscordMessageBotReset(discordMessageId string, msgHash string, larkMsg
 		MsgHash:      msgHash,
 	})
 	if err != nil {
-		chore.ReplyMsg(context.Background(), fmt.Sprintf("🤖️：图片重新生成失败，请稍后再试~\n错误信息: %v", err), &larkMsgId)
+		chore.ReplyMsg(
+			context.Background(),
+			fmt.Sprintf("🤖️：图片重新生成失败，请稍后再试~\n错误信息: %v", err),
+			&larkMsgId,
+		)
 		return err
 	}
-	chore.ReplyMsg(context.Background(), "🤖️：图片正在进行重新生成，请稍等......", &larkMsgId)
+	chore.ReplyMsg(
+		context.Background(),
+		"🤖️：图片重新生成中，请稍等......",
+		&larkMsgId,
+	)
 	return nil
 }
 
@@ -254,6 +281,7 @@ func UpscaleEventType(str string) string {
 }
 
 func instructException(id string, larkChatId string, msgId string) {
+	// 不一定是指令异常 也有可能是任务过多阻塞了
 	chore.ReplyMsg(context.Background(), "🤖️ ：发送的指令存在异常，请检查后重试", &msgId)
 	discordQueueDel(larkChatId)
 	db.GetCache().Clear(id)
@@ -272,19 +300,23 @@ func getDiscordLardMapId(content string) (id string, err error) {
 func getDiscordLarkMapJson(id string) (IDiscordLarkMap, error) {
 	discordLark := db.GetCache().GetInterface(id)
 	if discordLark == nil {
-		fmt.Println("get discord lark map json: ", id, "not found")
 		return IDiscordLarkMap{}, errors.New("not found")
 	}
 
 	var idl IDiscordLarkMap
 	if err := json.Unmarshal(discordLark, &idl); err != nil {
-		fmt.Println("get discord lark map json: ", "** json.Unmarshal失败 **", err)
 		return IDiscordLarkMap{}, errors.New("not found")
 	}
 	return idl, nil
 }
 
-func discordTriggerReplayLark(url string, discordMessageId string, key string, msgHash string, referenceMsgId string) {
+func discordTriggerReplayLark(
+	url string,
+	discordMsgId string,
+	key string,
+	msgHash string,
+	referenceMsgId string,
+) {
 	idl, err := getDiscordLarkMapJson(key)
 	if err != nil {
 		return
@@ -294,10 +326,15 @@ func discordTriggerReplayLark(url string, discordMessageId string, key string, m
 	if referenceMsgId != "" {
 		msgId = idl.LarkMsgIdMapPrevDiscordMsgId[referenceMsgId]
 	}
-	/** 重置会话状态 **/
 	discordQueueDel(idl.LarkChatId)
-	/** 回复飞书 **/
-	chore.ReplayImageByImagesDiscord(url, key, discordMessageId, msgHash, msgId, idl.From == "U")
+	chore.ReplayImageByImagesDiscord(
+		url,
+		key,
+		discordMsgId,
+		msgHash,
+		msgId,
+		idl.From == "U",
+	)
 }
 
 func discordIteratorTag(key string) {
@@ -313,100 +350,4 @@ func discordIteratorTag(key string) {
 func generateDiscordMsgHash(url string) string {
 	_parts := strings.Split(url, "_")
 	return strings.Split(_parts[len(_parts)-1], ".")[0]
-}
-
-type IDiscordQueue struct {
-	LarkChatId string `json:"larkChatId"`
-	Time       int64  `json:"time"`
-}
-
-func discordQueueCheck(larkChatId string) error {
-	for {
-		isLock := db.GetCache().Get(DiscordLockKey)
-		if isLock == "" {
-			break
-		}
-		time.Sleep(1 * time.Second)
-	}
-	db.GetCache().SetCustom(DiscordLockKey, "lock", time.Duration(2)*time.Second)
-	defer db.GetCache().Clear(DiscordLockKey)
-
-	// 下面是正常逻辑
-	queue := db.GetCache().GetInterface(DiscordQueueKey)
-
-	if queue != nil {
-		var queueList []IDiscordQueue
-		if err := json.Unmarshal(queue, &queueList); err != nil {
-			fmt.Println("discord queue check: ", "** json.Unmarshal失败 **", err)
-			return nil
-		}
-		queueList = discordQueueAutoOutDie(queueList)
-		db.GetCache().SetInterfaceNotTimeLimit(DiscordQueueKey, queueList)
-
-		for _, item := range queueList {
-			if item.LarkChatId == larkChatId {
-				return errors.New("🤖️：您存在任务正在运行中，请稍后再试～")
-			}
-		}
-
-		if len(queueList) > 3 {
-			return errors.New("🤖️：当前任务过多，请稍后再试～")
-		}
-	}
-	return nil
-}
-
-func discordQueueSet(larkChatId string) {
-	queue := db.GetCache().GetInterface(DiscordQueueKey)
-
-	if queue != nil {
-		var queueList []IDiscordQueue
-		if err := json.Unmarshal(queue, &queueList); err != nil {
-			fmt.Println("discord queue set: ", "** json.Unmarshal失败 **", err)
-			return
-		}
-		queueList = discordQueueAutoOutDie(queueList)
-		queueList = append(queueList, IDiscordQueue{
-			LarkChatId: larkChatId,
-			Time:       time.Now().Unix(),
-		})
-		db.GetCache().SetInterfaceNotTimeLimit(DiscordQueueKey, queueList)
-	} else {
-		db.GetCache().SetInterface(DiscordQueueKey, []IDiscordQueue{{
-			LarkChatId: larkChatId,
-			Time:       time.Now().Unix(),
-		}})
-	}
-}
-
-func discordQueueDel(larkChatId string) {
-	queue := db.GetCache().GetInterface(DiscordQueueKey)
-
-	if queue != nil {
-		var queueList []IDiscordQueue
-		if err := json.Unmarshal(queue, &queueList); err != nil {
-			fmt.Println("discord queue del: ", "** json.Unmarshal失败 **", err)
-			return
-		}
-		queueList = discordQueueAutoOutDie(queueList)
-		newQueueList := make([]IDiscordQueue, 0)
-		for _, item := range queueList {
-			if item.LarkChatId != larkChatId {
-				newQueueList = append(newQueueList, item)
-			}
-		}
-		db.GetCache().SetInterfaceNotTimeLimit(DiscordQueueKey, newQueueList)
-	}
-}
-
-func discordQueueAutoOutDie(queueList []IDiscordQueue) []IDiscordQueue {
-	currentTime := time.Now().Unix()
-	newQueueList := make([]IDiscordQueue, 0)
-	for _, item := range queueList {
-		if item.Time+30*60 > currentTime {
-			newQueueList = append(newQueueList, item)
-		}
-	}
-
-	return newQueueList
 }
